Complete reimplementation of the diffusion solver using FDM.
The diffusion solver now it implemented with Finite Difference Method (at least until next reimplementation!), using ghost cell method for the Dirichlet boundary condition (Neumann is defined in the boundary naturally).
Domain class takes care of geometrical description. Several implementations of power iteration implement the eigenvalue solution (which is normalized to solution norm and has to be renormalized to obtain some absolute value). The diffusion solver is capable of slab, cylindrical and spherical geometry with uniform or non-uniform grids and heterogeneous material, implementing the full divergent of the current, instead of laplacian of flux.
It is also capable of handling lists of the material class, inherited from the previous solver and cross-sections defined in functions (as shown in examples).

Needs a clean up of both old code and consolidation of source driven and eigenvalue solvers.
Also needs to implement more analytical verifications.